Fix client bootstrap crash: move sdSignalPropagation.js out of entities/
index.js's client-manifest generator (LoadScriptsFromFolder) does a blind
fs.readdirSync over game/entities/ and treats every file there as a
self-registering entity class, naming it after its default export's own
.name rather than the filename. sdSignalPropagation's default export is a
plain function named PropagateSignal, so it got registered as
"entities/PropagateSignal" and the client tried to dynamically import a
nonexistent entities/PropagateSignal.js - failing the Promise.all over all
entity imports and hanging the entire client on load.

Non-entity shared modules in this codebase live at game/ root (sdSound.js,
sdKeyStates.js, sdWorld.js), which isn't auto-scanned. Moved
sdSignalPropagation.js there and fixed its own relative imports plus the
two importers (sdButton.js, sdNode.js) accordingly. Verified by loading the
page against the redeployed test server: main menu now renders instead of
hanging on an unhandled import rejection.
